Extract breadcrumbs to their own concern/mixin
Quite a lot of funtionality and could mixin
to just those controllers that need it.
